# docs/api-docs.yaml
openapi: 3.0.1
info:
  title: Tasa-API
  description: "API for synchronize TASA app data."
  version: "1.0"
servers:
  - url: http://localhost:8080/api
...
paths:
  /user/register:
    post:
      tags:
        - Users
      summary: Register a new user
      requestBody:
        description: Register a new user given a username and password
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegisterInput'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Username too short
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '409':
          description: A user with such username already exists
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /user/login:
    post:
      tags:
        - Users
      summary: Retrieve a id, a token and a username given a username and a password, a login operation
      requestBody:
        description: Username and password of user
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginCredentialsInput'
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginOutput'
        '400':
          description: Username missing or invalid
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: A user with such username does not exists
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /user/logout:
    post:
      tags:
        - Users
      security:
        - bearerAuth: [ ]
      summary: Logout a user
      responses:
        '200':
          description: User details
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: A user with such username does not exists
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /user/{id}:
    get:
      security:
        - bearerAuth: []
      tags:
        - Users
      summary: Get details of a user
      description: Returns a single user
      parameters:
        - name: id
          in: path
          description: User identifier
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Username missing or invalid
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: User not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    User:
      type: object
      description: User details
      required:
        - id
        - username
        - email
      properties:
        id:
          type: integer
          description: user identifier
        username:
          type: string
          description: user username
          example: Bob
        email:
          type: string
          description: user email
          example: bob@example.pt
    UserRegisterInput:
      type: object
      description: User details for register
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
          description: User username
          example: Bob
        email:
          type: string
          description: User email
          example: bob@example.com
        password:
          type: string
          description: User password
          example: Strong_but_known_password1234
    TokenExternalInfo:
      type: object
      required:
        - expiration
        - refreshToken
        - token
      properties:
        token:
          type: string
          description: session access token
          example: absbadbdaksjdad_aksdjakld1223
        refreshToken:
          type: string
          description: session refresh token
        expiration:
          type: string
          format: date-time
          description: Time the session expires
          example: "2024-10-01T12:00:00Z"
...
    LoginOutput:
      type: object
      required:
        - session
        - user
      properties:
        user:
          $ref: "#/components/schemas/User"
        session:
          $ref: "#/components/schemas/TokenExternalInfo"
    UserLoginCredentialsInput:
      type: object
      description: User credentials for login
      required:
        - username
        - password
      properties:
        username:
          type: string
          description: user username
          example: Bob
        password:
          type: string
          description: user password
          example: Strong_but_known_password1234
...
    Problem:
      description: Error response
      type: object
      properties:
        type:
          type: string
        title:
          type: string
          format: uri
        status:
          type: integer
        detail:
          type: string